fix(drivers): watch API requries calling get and re-authentication (#297)
The correct way of calling the `watch` API is to first make a `list` call
then use the `ResourceVersion` as a starting point to stream the events.
Also, the API client credential may need re-authentication for longer
watching processes.

The existing `waitForJob` method lacks the two things, wrongly assuming
that a single `watch` call will handle the whole watching process. We
get errors like

 * Not handling `nil` `evt.Object` (end of a watch call)
 * Getting code `410 Gone` due to old `ResourceVersion`
 * Auth errors due to expired credential

This commit fixes all these.

The preferred way of handling waiting is to use `informer` APIs. An
attempt has beenn made in PR #296, but unfortunately it has to be
abandoned for now due to some shortcomings. We will probably re-visit
the `informer` API after a few months.