test(api): using io.Pipe to mock eventbus
The old way of using `bytes.Buffer` to mock the `eventbus` doesn't block
when there is no data available, the test code has to have a logic to
loop and check to wait for the messages, thus creating timing issue. The
new way of using `io.Pipe` mimics the actual `eventbus` behavior in blocking.

 * switching to `io.Pipe`
 * increasing the allowed delay for waiting