test(api): avoid timing issue injecting api messages
The test code injects the api ack/reply messages as soon as the outgoing request is sent. However
the old way signalling seems unreliable. It sometimes takes longer than 10ms between closing a channel and
receiving a zero value from the channel on the other end. The new code just put the starting of
message injection and completion of message sending together so no signalling is needed. This
should avoid the timing issue.