fix(config): avoid concurrent map write in config
The concurrent map writes are usually caused by updating to
the configuration when the services decrypt some piece of
data, or a regexp is parsed.

This commit changes how the configuration handles the updating
of the configs. Instead of, delaying the decryption and regexp
parsing to later time, the config module will take care of them
at the startup of the services.

The config process will be divided into multiple stages, and
all the services will be synchronized to use the configuration
according to the stages. Also introduced a `RawData` to holds
the raw data for use in rollback and re-decrypt, if necessary.
The configuration becomes readonly to services.

 * Add stages and the synchronizing variables to config
 * Put decryption and regexp parsing into stages
 * Remove scattered calls for decryption and regexp parsing
 * Add a `RawData` in config to keep `Rollback` working
 * Fixing tests accordingly