fix(services): api service discovery using staged dataset (#369)
* fix(services): api service discovery using staged dataset

The `APIFeatures` method is in charge of discovering what drivers to
load for api service. The old way is to use dataset in `APICfg`.  With
the new staged config loading method since version 2.1.3, it doesnt work
any more because the `APICfg` is not initialized during discovery. It is
supposed to use the dataset object passed in as parameters.

* test(integration): add api related tests in integration test

add api related tests in integration test