fix(dipper): driver ReadySignal chan initialization
A `chan` needs to be initialized before first use. We are missing the
initialization part for `ReadySignal`, but for some reason `go` does not
check for it, and allow you to send to or receive from `0x0`, and
potentially corrupting memmory.

The `ReadySignal` is used for signalling the readiness of the
configuration received from the `daemon`. Originally, it was considered
to be optional, since not all driver requires a configuration from
`daemon`. So, a timeout of one second was in place to forcefully start
the driver even without the signal.  I am changing this to be mandatory
so it can be a little easier sequencing the lifecycle, and avoid race
conditions. The `Driver` is receiving an empty configuration from
`daemon` even if no configuation is needed.