fix(config): contaminated dataset crashes during reload (#393)
Each repo in config keeps its own raw DataSet besides the final
assembled DataSet. When changes are detetcted, new and unchanged repo
DataSets are used for re-assembling the updated final DataSet. Some of
the changes that made to the final assembled DataSet spills into the raw
repo DataSet, and makes it unsuitable for reuse, and in turn causes
re-assembling to crash during reload. To fix this,

 - Make deep copies to avoid contaminating the original copy
 - Adjust the locking and reload detection in AdvanceStage