fix: API default timeout use store writeTimeout (#481)
* test: type match API testing with longer timeout

This is the only test seems still constantly timing out for the last couple of
months. Increasing the actTimeout should improve this as it did for the type all test.
The better resource class since last PR also helps.

* fix: API default timeout use store writeTimeout

When building an API request, the default timeout should match the store
configured `writeTimeout`. When using the constant value as default, it causes
the API has a longer timeout, and confuses the store yielding 202s instead of
500 timeouts.