fix: commandHandler protecting the return labels (#519)
* fix: commandHandler protecting the return labels

commandHandler uses the source msg labels to create the return msg, so if the labels are
updated in the function call, the return message will have dirty labels. This causes function
returns get send to a wrong recipient in cluster environments. An example is when a workflow
calls broadcast method of the redispubsub driver, the source labels were reused to send the
broadcast message, and `from` label was updated to local node IP. The function never returns
to the caller. By making a copy in the commandHandler, this can be avoid.

* refact: shorten commandHandler router func with commandWrapper