refactor: abstract comm I/O with MessageReceiver  (#528)
* refact: abstract comm I/O with MessageReceiver

Direct I/O is hidden so that comm API can be easily mocked, tested and
extended in the future.

* refact: driver runtime should be abstract over builtin driver

Driver is made up with an abstract runtime and a handler for each driver
type. The runtime should be abstract, so non-abstract features should be
moved to the handler.

* test: correct test after comm and driver change

Using NullReceiver to mock the driver I/O, and using NullDriverHandler
to mock driver functionality.