feat: support negated condition in match conditions
Use `!` suffix in variable name to negate a condition. Use `:except:` to
negate a combination of conditions in a map. Add `:present:` for checking
existence of a key in a map.

Negating a condition is useful since Golang regular expression doesn't
support negative lookahead. Sometime it is needed to negate parts of the
regex condition.