feat: cli job mode (#601)
* feat: cli job mode

Allow running the honeydipper daemon in job mode. Meaning it will load a file specified by `JOB_FILE` run a workflow named `reserved/main` and quit once complete. Useful for test/debug locally. Will add the document for this in a separate PR.

* fix: use main branch by default

In addition to #590, a few more places need to change to allow using main as default branch.

* fix: linter errors

* fix: StartSession should block until session is persisted

* fix: skip periodical config refresh for job mode

* fix: correct timing for start the job main workflow