feat: support hide thinking messages for reasoning models (#647)
* refactor: use the ai package for ollama driver

The ai package contains the logic that can be shared among different AI
model APIs. It is tested, abstract and independent. It makes sense to
take advantage of this framework.

* feat: support hide thinking messages for reasoning models

Reasoning models produce thinking messages, and if hide_think is set to
true. These messages will not show up in the chat, neither will they be
stored in the chat history, making the subsequential chat reqeusts
easier to transmit.

* doc: add a FAQ

* fix: deserialize the tool call payload before use